+++ android/src/main/kotlin/com/dc16/wechat_video_call/WeChatActivity.kt
@@ -1,6 +1,6 @@
 package com.dc16.wechat_video_call
 
-enum class WechatActivity(val id: String) {
+enum class WeChatActivity(val id: String) {
     INDEX("com.tencent.mm.ui.LauncherUI"),
     CHAT("com.tencent.mm.ui.chatting.ChattingUI"),
     SEARCH("com.tencent.mm.plugin.fts.ui.FTSMainUI"),