+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-## 0.0.2
+## 1.0.0
+* Fix: requestAccessibilityPermission is no return value
+* Fix: WeChat spelling
+* Optimize WeChat return to home page
 
+## 0.0.3
+* Configure automated publishing through GitHub Actions
+
+## 0.0.2
 * Add support Android WeChat 8.0.51 (2720).
-* code formatting
+* Code formatting
 * Support voice call
 
 ## 0.0.1
-
 * Initial implementation of functions, support Android WeChat 8.0.50 (2701).